+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 - `limestone/Card` prop `imageSize` to set the size of card image
 - `limestone/Icon` supported icon list, adding new icons
+- `limestone/VideoPlayer` prop `includeTimeHour` to conditionally show hour for current and total time
 
 ### Changed